requirements.txt
telethon
speedtest
toml
gitpython
